chore: refactor the blocklist component folder
