+++ packages/twenty-front/src/modules/settings/accounts/components/__stories__/SettingsAccountsBlocklistInput.stories.tsx
@@ -2,7 +2,7 @@ import { Decorator, Meta, StoryObj } from '@storybook/react';
 import { expect, fn, userEvent, within } from '@storybook/test';
 import { ComponentDecorator } from 'twenty-ui';
 
-import { SettingsAccountsBlocklistInput } from '@/settings/accounts/components/blocklist/SettingsAccountsBlocklistInput';
+import { SettingsAccountsBlocklistInput } from '@/settings/accounts/components/blocklist/components/SettingsAccountsBlocklistInput';
 
 const updateBlockedEmailListJestFn = fn();
