+++ packages/twenty-front/src/modules/settings/accounts/components/__stories__/SettingsAccountsBlocklistSection.stories.tsx
@@ -1,8 +1,8 @@
 import { Meta, StoryObj } from '@storybook/react';
 import { ComponentDecorator } from 'twenty-ui';
 
-import { SettingsAccountsBlocklistInput } from '@/settings/accounts/components/blocklist/SettingsAccountsBlocklistInput';
-import { SettingsAccountsBlocklistSection } from '@/settings/accounts/components/blocklist/SettingsAccountsBlocklistSection';
+import { SettingsAccountsBlocklistInput } from '@/settings/accounts/components/blocklist/components/SettingsAccountsBlocklistInput';
+import { SettingsAccountsBlocklistSection } from '@/settings/accounts/components/blocklist/components/SettingsAccountsBlocklistSection';
 
 const meta: Meta<typeof SettingsAccountsBlocklistSection> = {
   title: 'Modules/Settings/Accounts/Blocklist/SettingsAccountsBlocklistSection',
